Title: 10개 입력받기, Time: 6ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240820/10개 입력받기/receive-10-inputs.c
@@ -13,7 +13,7 @@ int main() {
         sum_value += arr[i];
         cnt++;
     }
-    avg = sum_value/cnt;
+    avg = (double)sum_value/cnt;
     printf("%d %.1lf",sum_value,avg);
     return 0;
 }
